Add working EnforceAuth Middleware Placeholdre
