+++ Cargo.toml
@@ -27,6 +27,7 @@ bytes = "0.4"
 actix-web = "1.0.7"
 actix-files = "0.1.4"
 actix-http = "0.2.9"
+actix-service = "0.4.2"
 geojson = "0.16.0"
 quick-xml = "0.16.1"
 serde = "1.0"
